Remove unwanted imports
